activities: Set nextPageUrl to null when manually refreshing activities.
When the user swipes the recycler view to refresh the list of activities we want the list to be cleared and populated with the newest activities without loading all the other paginated activities that might have been loaded up to that point. Therefore we set the nextPageUrl variable to null in the setOnRefreshListener for the swipeRefreshLayout.