build: Exclude xpp3 dependency from library
Although this dep is excluded in the library itself, Gradle still
includes it because it doesn't respect exclusions for transitive dependencies.
(see https://github.com/gradle/gradle/issues/1473)

This would be okay if it wasn't for the fact that xpp3 forces an old Junit version
that breaks the integration tests.

Signed-off-by: Álvaro Brey Vilas <alvaro.brey@nextcloud.com>