code style: Disable autowrap while typing
Our .editorconfig indiscriminately applies max length of 120 to every single file (except yaml).
When coupled with this setting (autowrap), this results in the editor inserting line ends in files,
even when it breaks the syntax (such as bash scripts) or tools (such as strings.xml).

This change keeps the margin indicator but disables auto-wrap while typing.

Signed-off-by: Álvaro Brey Vilas <alvaro.brey@nextcloud.com>