Added the gradle signing key to the keyring to fix project sync error
The commit 98a9934ae4f74dfba3e3f5cecb0443ea0872afb0 introduced stricter
dependency verification, but this resulted in the Gradle project syncing
to fail with the error:
> Multiple build operations failed.
> Dependency verification failed for configuration ':app:androidTestUtil':
>   - On artifact gradle-9.2.1-src.zip (gradle:gradle:9.2.1) in repository
> 'Gradle distributions': Artifact was signed with key 'E2F38302C8075E3D'
> but it wasn't found in any key server so it couldn't be verified
> This can indicate that a dependency has been compromised. Please
> carefully verify the signatures and checksums

This happened inpedendent of the used gradle version..

The followup 19da59e8b8014b95323a01e85501da3e04aa9fb8 added the key's
fingerprint, but this was not enough to establish trust. Adding the public
key to the keyring. The key was taken from
https://github.com/gradle/gradle/blob/b99fe7d304e68b978011840c31c08068f671f794/KEYS.md
and cross-checked with https://gradle.org/keys/

Signed-off-by: Philipp Hasper <vcs@hasper.info>